Move hybridcompute/2019-12-12 out of preview since it should be stable
