+++ CHANGELOG.md
@@ -49,6 +49,8 @@
 
 - package `github.com/Azure/azure-sdk-for-go/services/servicefabric/mgmt/2018-02-01/servicefabric` has been renamed to `github.com/Azure/azure-sdk-for-go/services/preview/servicefabric/mgmt/2018-02-01-preview/servicefabric`
 
+- package `github.com/Azure/azure-sdk-for-go/services/preview/hybridcompute/mgmt/2019-12-12/hybridcompute` has been renamed to `github.com/Azure/azure-sdk-for-go/services/hybridcompute/mgmt/2019-12-12/hybridcompute`
+
 Generated from https://github.com/Azure/azure-rest-api-specs/tree/93106ff722d481c30a9a4ae089564635c10e9401
 
 ## `v46.4.0`
