+++ profiles/preview/preview/hybridcompute/mgmt/hybridcompute/models.go
@@ -22,7 +22,7 @@ package hybridcompute
 import (
 	"context"
 
-	original "github.com/Azure/azure-sdk-for-go/services/preview/hybridcompute/mgmt/2019-12-12/hybridcompute"
+	original "github.com/Azure/azure-sdk-for-go/services/preview/hybridcompute/mgmt/2019-03-18-preview/hybridcompute"
 )
 
 const (
@@ -35,107 +35,38 @@ const (
 	InstanceView InstanceViewTypes = original.InstanceView
 )
 
-type ResourceIdentityType = original.ResourceIdentityType
-
-const (
-	SystemAssigned ResourceIdentityType = original.SystemAssigned
-)
-
-type SkuTier = original.SkuTier
-
-const (
-	Basic    SkuTier = original.Basic
-	Free     SkuTier = original.Free
-	Premium  SkuTier = original.Premium
-	Standard SkuTier = original.Standard
-)
-
-type StatusLevelTypes = original.StatusLevelTypes
-
-const (
-	Error   StatusLevelTypes = original.Error
-	Info    StatusLevelTypes = original.Info
-	Warning StatusLevelTypes = original.Warning
-)
-
 type StatusTypes = original.StatusTypes
 
 const (
-	StatusTypesConnected    StatusTypes = original.StatusTypesConnected
-	StatusTypesDisconnected StatusTypes = original.StatusTypesDisconnected
-	StatusTypesError        StatusTypes = original.StatusTypesError
+	Connected    StatusTypes = original.Connected
+	Disconnected StatusTypes = original.Disconnected
+	Error        StatusTypes = original.Error
 )
 
-type AzureEntityResource = original.AzureEntityResource
 type BaseClient = original.BaseClient
-type ErrorAdditionalInfo = original.ErrorAdditionalInfo
 type ErrorDetail = original.ErrorDetail
 type ErrorResponse = original.ErrorResponse
 type Identity = original.Identity
-type LocationData = original.LocationData
 type Machine = original.Machine
-type MachineExtension = original.MachineExtension
-type MachineExtensionInstanceView = original.MachineExtensionInstanceView
-type MachineExtensionInstanceViewStatus = original.MachineExtensionInstanceViewStatus
-type MachineExtensionProperties = original.MachineExtensionProperties
-type MachineExtensionPropertiesInstanceView = original.MachineExtensionPropertiesInstanceView
-type MachineExtensionPropertiesModel = original.MachineExtensionPropertiesModel
-type MachineExtensionUpdate = original.MachineExtensionUpdate
-type MachineExtensionUpdateProperties = original.MachineExtensionUpdateProperties
-type MachineExtensionUpdatePropertiesModel = original.MachineExtensionUpdatePropertiesModel
-type MachineExtensionsClient = original.MachineExtensionsClient
-type MachineExtensionsCreateOrUpdateFuture = original.MachineExtensionsCreateOrUpdateFuture
-type MachineExtensionsDeleteFuture = original.MachineExtensionsDeleteFuture
-type MachineExtensionsListResult = original.MachineExtensionsListResult
-type MachineExtensionsListResultIterator = original.MachineExtensionsListResultIterator
-type MachineExtensionsListResultPage = original.MachineExtensionsListResultPage
-type MachineExtensionsUpdateFuture = original.MachineExtensionsUpdateFuture
-type MachineIdentity = original.MachineIdentity
 type MachineListResult = original.MachineListResult
 type MachineListResultIterator = original.MachineListResultIterator
 type MachineListResultPage = original.MachineListResultPage
 type MachineProperties = original.MachineProperties
-type MachinePropertiesModel = original.MachinePropertiesModel
-type MachinePropertiesOsProfile = original.MachinePropertiesOsProfile
 type MachineReconnect = original.MachineReconnect
 type MachineReconnectProperties = original.MachineReconnectProperties
-type MachineReconnectPropertiesModel = original.MachineReconnectPropertiesModel
 type MachineUpdate = original.MachineUpdate
-type MachineUpdateIdentity = original.MachineUpdateIdentity
 type MachineUpdateProperties = original.MachineUpdateProperties
-type MachineUpdatePropertiesModel = original.MachineUpdatePropertiesModel
 type MachinesClient = original.MachinesClient
 type OSProfile = original.OSProfile
 type OperationListResult = original.OperationListResult
 type OperationValue = original.OperationValue
 type OperationValueDisplay = original.OperationValueDisplay
-type OperationValueDisplayModel = original.OperationValueDisplayModel
 type OperationsClient = original.OperationsClient
-type Plan = original.Plan
-type ProxyResource = original.ProxyResource
 type Resource = original.Resource
-type ResourceModelWithAllowedPropertySet = original.ResourceModelWithAllowedPropertySet
-type ResourceModelWithAllowedPropertySetIdentity = original.ResourceModelWithAllowedPropertySetIdentity
-type ResourceModelWithAllowedPropertySetPlan = original.ResourceModelWithAllowedPropertySetPlan
-type ResourceModelWithAllowedPropertySetSku = original.ResourceModelWithAllowedPropertySetSku
-type Sku = original.Sku
-type TrackedResource = original.TrackedResource
 type UpdateResource = original.UpdateResource
 
 func New(subscriptionID string) BaseClient {
 	return original.New(subscriptionID)
-}
-func NewMachineExtensionsClient(subscriptionID string) MachineExtensionsClient {
-	return original.NewMachineExtensionsClient(subscriptionID)
-}
-func NewMachineExtensionsClientWithBaseURI(baseURI string, subscriptionID string) MachineExtensionsClient {
-	return original.NewMachineExtensionsClientWithBaseURI(baseURI, subscriptionID)
-}
-func NewMachineExtensionsListResultIterator(page MachineExtensionsListResultPage) MachineExtensionsListResultIterator {
-	return original.NewMachineExtensionsListResultIterator(page)
-}
-func NewMachineExtensionsListResultPage(getNextPage func(context.Context, MachineExtensionsListResult) (MachineExtensionsListResult, error)) MachineExtensionsListResultPage {
-	return original.NewMachineExtensionsListResultPage(getNextPage)
 }
 func NewMachineListResultIterator(page MachineListResultPage) MachineListResultIterator {
 	return original.NewMachineListResultIterator(page)
@@ -160,15 +91,6 @@ func NewWithBaseURI(baseURI string, subscriptionID string) BaseClient {
 }
 func PossibleInstanceViewTypesValues() []InstanceViewTypes {
 	return original.PossibleInstanceViewTypesValues()
-}
-func PossibleResourceIdentityTypeValues() []ResourceIdentityType {
-	return original.PossibleResourceIdentityTypeValues()
-}
-func PossibleSkuTierValues() []SkuTier {
-	return original.PossibleSkuTierValues()
-}
-func PossibleStatusLevelTypesValues() []StatusLevelTypes {
-	return original.PossibleStatusLevelTypesValues()
 }
 func PossibleStatusTypesValues() []StatusTypes {
 	return original.PossibleStatusTypesValues()
