+++ services/hybridcompute/mgmt/2019-12-12/hybridcompute/enums.go
