+++ services/hybridcompute/mgmt/2019-12-12/hybridcompute/machines.go
